FEATURE: support serialization of query results
Previous to this change there was no clean way of marshalling
results.

This change introduces:

```
r = MiniSql::Serializer.to_json(DB.query("select 1 A"))
MiniSql::Serializer.from_json(r).first.A
```

I opted for JSON here as it is slightly more portable than
Marshal

Performance should be good, but we can improve by adding a
to_a implementation to all query results so we can avoid
serializing the keys for every row.